web/requirements.txt
celery==5.1.2
django==3.2.5
django-crispy-forms==1.12.0
django-widget-tweaks==1.4.8
pillow==8.3.1
pip-chill==1.0.1
priority==1.3.0
psycopg2-binary==2.9.1
python-decouple==3.4
python-dotenv==0.19.0
python-scrapyd-api==2.1.2
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.2.0a1
virtualenv==20.8.1
Scrapy==2.5.0
pytz==2021.3